requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.1.0
black==19.3b0
chardet==3.0.4
Click==7.0
discord.py==1.2.3
idna==2.8
multidict==4.5.2
toml==0.10.0
websockets==6.0
yarl==1.3.0
